Switch from sunxi to upstream U-Boot
The patches we need were merged long ago.

Also, fix the Xen branch we use (was lost when the repository became a
fork of mirage/xen).

The new U-Boot has a different memory layout and so requires Xen and
Linux to be loaded at a different address. Thanks to Luke Dunstan for
finding out what had moved there and to Robbie VanVossen for updating
for the CubieTruck.

Fixes #57.